+++ src/textureminer/exceptions.py
